1.0.0-alpha.4 - moved shader dataloader location due to 1.21.2 shader changes. Shaders are now loaded from `namespace:luminance/shader_name.json`.

+++ src/main/java/com/mclegoman/luminance/mixin/client/shaders/GameRendererAccessor.java
@@ -15,5 +15,5 @@ import org.spongepowered.asm.mixin.gen.Invoker;
 @Mixin(GameRenderer.class)
 public interface GameRendererAccessor {
 	@Invoker("getFov")
-	double invokeGetFov(Camera camera, float tickDelta, boolean changingFov);
+	float invokeGetFov(Camera camera, float tickDelta, boolean changingFov);
 }
